try to FIX #102

+++ native/common/jp_javaenv.cpp
@@ -140,7 +140,7 @@ int JPJavaEnv::DestroyJavaVM()
 void JPJavaEnv::DeleteLocalRef(jobject obj)
 {
 	TRACE_IN("JPJavaEnv::DeleteLocalRef");
-	TRACE1((long)obj);
+	TRACE1(obj);
 	JNIEnv* env = getJNIEnv();
 	if (env != NULL)
 	{
@@ -152,7 +152,7 @@ void JPJavaEnv::DeleteLocalRef(jobject obj)
 void JPJavaEnv::DeleteGlobalRef(jobject obj)
 {
 	TRACE_IN("JPJavaEnv::DeleteGlobalRef");
-	TRACE1((long)obj);
+	TRACE1(obj);
 	JNIEnv* env = getJNIEnv();
 	if (env != NULL)
 	{
@@ -164,11 +164,11 @@ void JPJavaEnv::DeleteGlobalRef(jobject obj)
 jobject JPJavaEnv::NewLocalRef(jobject a0)
 {
 	TRACE_IN("JPJavaEnv::NewLocalRef");
-	TRACE1((long)a0);
+	TRACE1(a0);
 	jobject res;
 	JNIEnv* env = getJNIEnv();
 	res = env->functions->NewLocalRef(env, a0);
-	TRACE1((long)res); //, JPJni::getClassName(a0).getSimpleName());
+	TRACE1(res); //, JPJni::getClassName(a0).getSimpleName());
 	return res;
 	TRACE_OUT;
 }
@@ -176,11 +176,11 @@ jobject JPJavaEnv::NewLocalRef(jobject a0)
 jobject JPJavaEnv::NewGlobalRef(jobject a0)
 {
 	TRACE_IN("JPJavaEnv::NewGlobalRef");
-	TRACE1((long)a0);
+	TRACE1(a0);
 	jobject res;
 	JNIEnv* env = getJNIEnv();
 	res = env->functions->NewGlobalRef(env, a0);
-	TRACE1((long)res); //, JPJni::getClassName(a0).getSimpleName());
+	TRACE1(res); //, JPJni::getClassName(a0).getSimpleName());
 	return res;
 	TRACE_OUT;
 }
@@ -284,7 +284,7 @@ jobject JPJavaEnv::NewDirectByteBuffer(void* address, jlong capacity)
 	JNIEnv* env = getJNIEnv(); 
     jobject res = env->functions->NewDirectByteBuffer(env, address, capacity);
     JAVA_CHECK("NewDirectByteBuffer");
-	TRACE1((long)res);
+	TRACE1(res);
     return res;	
 	TRACE_OUT;
 }
